+++ CHANGES.md
@@ -3,6 +3,9 @@
 - Fix exception `__cause__` not being propagated with
   `tblib.pickling_support.install()` (#255).
 
+- Fix handling of CPU affinity  by using `psutil`'s `cpu_affinity` on platforms
+  that do not implement `os.sched_getaffinity`, such as PyPy (#381).
+
 ### 3.3.0 - 2022-09-15
 
 - Fix worker management logic in `get_reusable_executor` to ensure